refactor: RepoIngestionService에서 트랜잭션 및 저장 로직 분리
- RepoIngestionService에서 API 호출과 저장 로직을 분리하여 트랜잭션 경계 최소화
  - API 호출 중 DB 커넥션 점유/락 발생 가능성 제거
  - 저장 로직은 RepoUpsertService로 이동하여 @Transactional 적용
- RepoUpsertService 추가
  - GitHub 계정 로딩 후 저장소 엔티티 변환 및 saveAll 처리
  - PK(repoId) 기반 upsert 수행
